+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+- 0.4.0.1 (2017-12-26)
+    * Bump `tasty` and `tasty-hunit` dependencies
+
 - 0.4.0.0 (2016-12-22)
     * Improved error messages
     * Added a simple test suite